Move IMGUI sdl implementation to backends/ folder
