+++ contrib/imgui/backends/imgui_impl_sdl2.cpp
@@ -51,7 +51,7 @@
 //  2016-10-15: Misc: Added a void* user_data parameter to Clipboard function handlers.
 
 #include "imgui.h"
-#include "imgui_impl_sdl.h"
+#include "imgui_impl_sdl2.h"
 
 // SDL
 #include <SDL.h>
